Safety and clarification
-bash script updated to avoid deleting things on accident if the script is called in a strange way
-note added to sp_meta.py to show the differences in metadata reader forks

+++ config/shairport_metadata.bash
@@ -1,5 +1,10 @@
 #!/bin/bash
 # Clear out any previous album cover images, then navigate to the proper directory #
+src="$1"
+if ((src < 0 || src >= 4));
+then
+    exit 1;
+fi
 rm -r -f /home/pi/web/static/imgs/srcs/$1/
 mkdir -p /home/pi/web/static/imgs/srcs/$1/
 cd /home/pi/web/static/imgs/srcs/$1/
